new extractors in progress
Former-commit-id: d541aaf569c3135955ea80dbc38ef4235cdea9af

+++ branches/new_data_structures/GenPlay/src/edu/yu/einstein/genplay/gui/dialog/trackSettings/ScoreOptionsPanel.java
@@ -205,16 +205,16 @@ class ScoreOptionsPanel extends JPanel implements ActionListener, ChangeListener
 	/**
 	 * @return the maximum score
 	 */
-	double getScoreMaximum() {
-		return ((Number) jftfScoreMax.getValue()).doubleValue();
+	float getScoreMaximum() {
+		return ((Number) jftfScoreMax.getValue()).floatValue();
 	}
 
 
 	/**
 	 * @return the minimun score
 	 */
-	double getScoreMinimum() {
-		return ((Number) jftfScoreMin.getValue()).doubleValue();
+	float getScoreMinimum() {
+		return ((Number) jftfScoreMin.getValue()).floatValue();
 	}
